+++ Repositories/Interfaces/iCityInfoReposit.cs
@@ -1,6 +1,5 @@
 ﻿using CityInfoApi.Dtos;
-using CityInfoApi.Models;
-using Microsoft.AspNetCore.Mvc;
+using CityInfoApi.Models_new;
 
 namespace CityInfoApi.Repositories.Interfaces
 {
@@ -21,7 +20,7 @@ namespace CityInfoApi.Repositories.Interfaces
 
         Task<IEnumerable<Sanad1400>> Getacc_SelectSanadsAsync(int code, string date1);
         Task<IEnumerable<Anbar>> GetAnbarsAsync();
-        Task<int> SubmitFactorAsync(FactorSubmit factorDetail);
+        Task<long> SubmitFactorAsync(FactorSubmit factorDetail);
         Task<Moshtari> CreateMoshtariAsync(MoshtariDto moshtariDto);
         Task<IEnumerable<GorohM>> GetGorohMsAsync();
     }
